Add IPV4Adress instead of ip : str Replace port with u16 instead of u32.
Close #33

+++ crates/crabe_io/src/league/game_controller/config.rs
@@ -7,5 +7,5 @@ pub struct GameControllerConfig {
     pub gc_ip: String,
 
     #[arg(long, default_value_t = 10020)]
-    pub gc_port: u32,
+    pub gc_port: u16,
 }
